+++ CHANGELOG.md
@@ -7,14 +7,19 @@
 
 ### Added
 *   Added `appear_on_main` field to `Categories` field that will allow for querying only categories that, well, we only want to appear on the main page.
+*   Added a new `website` endpoint that is more catered to interacting with React.
 
 ### Fixed
+*   Fixed the type checking for API responses to be `JsonResponse` and not `HttpResponse`.
 
 ### Changes
 *   Changed the API so it is separated into "general"/"standard" and "website" API requests.
     *   React will be using a lot of the thps.run API, so separating this will help keep features separate and also allow us to do fancier things.
 
 ### Removed
+*   (Planned) Removed the `subcategory` field from the `Runs` model.
+    *   This was a remnant from v2 code, before it was more understood how to link variable-value pairs to an individual run.
+*   (Planned) Removed the `place` field from the `Runs` model.
 
 ### Misc.